disease model/requirements.txt
torch
torchvision
tqdm
scikit-learn
Pillow
